open-ended provider version selection

+++ cbtom-cat/manifests/upgrade.pp
@@ -39,11 +39,13 @@ class tom_cat::upgrade (
 
     exec { 'stop_tomcat_before_upgrade_windows':
       command   => "${windows_powershell} -Command \"if (Get-Service -Name '${service_name}' -ErrorAction SilentlyContinue) { Stop-Service -Name '${service_name}' -Force }\"",
+      onlyif    => "${windows_powershell} -Command \"if (Get-Service -Name '${service_name}' -ErrorAction SilentlyContinue) { exit 0 } elseif (Test-Path '${windows_install_dir}') { exit 0 } else { exit 1 }\"",
       logoutput => true,
     }
 
     exec { 'backup_existing_tomcat_windows':
       command   => "${windows_powershell} -Command \"if (Test-Path '${windows_install_dir}') { Compress-Archive -Path '${windows_install_dir}\\*' -DestinationPath 'C:\\temp\\${service_name}-${timestamp}.zip' -Force }\"",
+      onlyif    => "${windows_powershell} -Command \"if (Test-Path '${windows_install_dir}') { exit 0 } else { exit 1 }\"",
       logoutput => true,
       require   => Exec['stop_tomcat_before_upgrade_windows'],
     }
